Use bigint for ss_ticket_number
Avoid overflow

+++ ddl-tpcds/bin_partitioned/store_sales.sql
@@ -13,7 +13,7 @@ create table store_sales
     ss_addr_sk                int,
     ss_store_sk               int,
     ss_promo_sk               int,
-    ss_ticket_number          int,
+    ss_ticket_number          bigint,
     ss_quantity               int,
     ss_wholesale_cost         float,
     ss_list_price             float,
